+++ go/vt/vtgate/executor_framework_test.go
@@ -334,14 +334,7 @@ func init() {
 	vindexes.Register("keyrange_lookuper_unique", newKeyRangeLookuperUnique)
 }
 
-type executorType bool
-
-const (
-	legacy           executorType = true
-	planAllTheThings executorType = false
-)
-
-func createExecutorEnvUsing(t executorType) (executor *Executor, sbc1, sbc2, sbclookup *sandboxconn.SandboxConn) {
+func createExecutorEnv() (executor *Executor, sbc1, sbc2, sbclookup *sandboxconn.SandboxConn) {
 	// Use legacy gateway until we can rewrite these tests to use new tabletgateway
 	*GatewayImplementation = GatewayImplementationDiscovery
 	cell := "aa"
@@ -369,22 +362,10 @@ func createExecutorEnvUsing(t executorType) (executor *Executor, sbc1, sbc2, sbc
 	bad.VSchema = badVSchema
 
 	getSandbox(KsTestUnsharded).VSchema = unshardedVSchema
-	switch t {
-	case legacy:
-		executor = NewExecutor(context.Background(), serv, cell, resolver, false, testBufferSize, testCacheSize)
-	case planAllTheThings:
-		f := func(executor *Executor) executeMethod {
-			return &planExecute{e: executor}
-		}
-		executor = NewTestExecutor(context.Background(), f, serv, cell, resolver, false, testBufferSize, testCacheSize)
-	}
+	executor = NewExecutor(context.Background(), serv, cell, resolver, false, testBufferSize, testCacheSize)
 
 	key.AnyShardPicker = DestinationAnyShardPickerFirstShard{}
 	return executor, sbc1, sbc2, sbclookup
-}
-
-func createExecutorEnv() (executor *Executor, sbc1, sbc2, sbclookup *sandboxconn.SandboxConn) {
-	return createExecutorEnvUsing(legacy)
 }
 
 func createCustomExecutor(vschema string) (executor *Executor, sbc1, sbc2, sbclookup *sandboxconn.SandboxConn) {
