+++ go/vt/vtgate/executor_test.go
@@ -27,7 +27,6 @@ import (
 	"reflect"
 	"strings"
 	"testing"
-	"time"
 
 	"vitess.io/vitess/go/test/utils"
 
@@ -51,37 +50,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestExecutorResultsExceeded(t *testing.T) {
-	save := *warnMemoryRows
-	*warnMemoryRows = 3
-	defer func() { *warnMemoryRows = save }()
-
-	executor, _, _, sbclookup := createExecutorEnv()
-	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master"})
-
-	initial := warnings.Counts()["ResultsExceeded"]
-
-	result1 := sqltypes.MakeTestResult(sqltypes.MakeTestFields("col", "int64"), "1")
-	result2 := sqltypes.MakeTestResult(sqltypes.MakeTestFields("col", "int64"), "1", "2", "3", "4")
-	sbclookup.SetResults([]*sqltypes.Result{result1, result2})
-
-	_, err := executor.Execute(context.Background(), "TestExecutorResultsExceeded", session, "select * from main1", nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := warnings.Counts()["ResultsExceeded"], initial; got != want {
-		t.Errorf("warnings count: %v, want %v", got, want)
-	}
-
-	_, err = executor.Execute(context.Background(), "TestExecutorResultsExceeded", session, "select * from main1", nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := warnings.Counts()["ResultsExceeded"], initial+1; got != want {
-		t.Errorf("warnings count: %v, want %v", got, want)
-	}
-}
 
 func TestExecutorTransactionsNoAutoCommit(t *testing.T) {
 	executor, _, _, sbclookup := createExecutorEnv()
@@ -1440,397 +1408,6 @@ func TestExecutorAddDropVschemaTableDDL(t *testing.T) {
 	}
 }
 
-func TestExecutorAddSequenceDDL(t *testing.T) {
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	defer func() {
-		*vschemaacl.AuthorizedDDLUsers = ""
-	}()
-	executor, _, _, _ := createExecutorEnv()
-	ks := KsTestUnsharded
-
-	vschema := executor.vm.GetCurrentSrvVschema()
-
-	var vschemaTables []string
-	for t := range vschema.Keyspaces[ks].Tables {
-		vschemaTables = append(vschemaTables, t)
-	}
-
-	session := NewSafeSession(&vtgatepb.Session{TargetString: ks})
-	stmt := "alter vschema add sequence test_seq"
-	_, err := executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	require.NoError(t, err)
-	_ = waitForVschemaTables(t, ks, append(vschemaTables, []string{"test_seq"}...), executor)
-	vschema = executor.vm.GetCurrentSrvVschema()
-	table := vschema.Keyspaces[ks].Tables["test_seq"]
-	wantType := "sequence"
-	if table.Type != wantType {
-		t.Errorf("want table type sequence got %v", table)
-	}
-
-	// Should fail adding a table on a sharded keyspace
-	ksSharded := "TestExecutor"
-	session = NewSafeSession(&vtgatepb.Session{TargetString: ksSharded})
-
-	stmt = "alter vschema add sequence sequence_table"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-
-	wantErr := "add sequence table: unsupported on sharded keyspace TestExecutor"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("want error %v got %v", wantErr, err)
-	}
-
-	// Should be able to add autoincrement to table in sharded keyspace
-	stmt = "alter vschema on test_table add vindex hash_index (id)"
-	if _, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil); err != nil {
-		t.Error(err)
-	}
-	time.Sleep(10 * time.Millisecond)
-
-	stmt = "alter vschema on test_table add auto_increment id using test_seq"
-	if _, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil); err != nil {
-		t.Error(err)
-	}
-	time.Sleep(10 * time.Millisecond)
-
-	wantAutoInc := &vschemapb.AutoIncrement{Column: "id", Sequence: "test_seq"}
-	gotAutoInc := executor.vm.GetCurrentSrvVschema().Keyspaces[ksSharded].Tables["test_table"].AutoIncrement
-
-	if !reflect.DeepEqual(wantAutoInc, gotAutoInc) {
-		t.Errorf("want autoinc %v, got autoinc %v", wantAutoInc, gotAutoInc)
-	}
-}
-
-func TestExecutorAddDropVindexDDL(t *testing.T) {
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	defer func() {
-		*vschemaacl.AuthorizedDDLUsers = ""
-	}()
-	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
-	ks := "TestExecutor"
-	session := NewSafeSession(&vtgatepb.Session{TargetString: ks})
-	vschemaUpdates := make(chan *vschemapb.SrvVSchema, 4)
-	executor.serv.WatchSrvVSchema(context.Background(), "aa", func(vschema *vschemapb.SrvVSchema, err error) {
-		vschemaUpdates <- vschema
-	})
-
-	vschema := <-vschemaUpdates
-	_, ok := vschema.Keyspaces[ks].Vindexes["test_hash"]
-	if ok {
-		t.Fatalf("test_hash should not exist in original vschema")
-	}
-
-	// Create a new vindex implicitly with the statement
-	stmt := "alter vschema on test add vindex test_hash (id) using hash "
-	_, err := executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	_, vindex := waitForVindex(t, ks, "test_hash", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	_ = waitForColVindexes(t, ks, "test", []string{"test_hash"}, executor)
-	qr, err := executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-	}
-	wantqr := &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-		},
-		RowsAffected: 1,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	// Drop it
-	stmt = "alter vschema on test drop vindex test_hash"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	_, _ = waitForVindex(t, ks, "test_hash", vschemaUpdates, executor)
-	_ = waitForColVindexes(t, ks, "test", []string{}, executor)
-	_, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	wantErr := "table `test` does not exist in keyspace `TestExecutor`"
-	if err == nil || err.Error() != wantErr {
-		t.Fatalf("expected error in show vschema vindexes on TestExecutor.test %v: got %v", wantErr, err)
-	}
-
-	// add it again using the same syntax
-	stmt = "alter vschema on test add vindex test_hash (id) using hash "
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	_, vindex = waitForVindex(t, ks, "test_hash", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	_ = waitForColVindexes(t, ks, "test", []string{"test_hash"}, executor)
-
-	qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-	}
-	wantqr = &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-		},
-		RowsAffected: 1,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	// add another
-	stmt = "alter vschema on test add vindex test_lookup (c1,c2) using lookup with owner=`test`, from=`c1,c2`, table=test_lookup, to=keyspace_id"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ks, "test_lookup", vschemaUpdates, executor)
-	if vindex.Type != "lookup" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	if table, ok := vschema.Keyspaces[ks].Tables["test"]; ok {
-		if len(table.ColumnVindexes) != 2 {
-			t.Fatalf("table vindexes want 1 got %d", len(table.ColumnVindexes))
-		}
-		if table.ColumnVindexes[1].Name != "test_lookup" {
-			t.Fatalf("table vindexes didn't contain test_lookup")
-		}
-	} else {
-		t.Fatalf("table test not defined in vschema")
-	}
-
-	qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-	}
-	wantqr = &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-			buildVarCharRow("c1, c2", "test_lookup", "lookup", "from=c1,c2; table=test_lookup; to=keyspace_id", "test"),
-		},
-		RowsAffected: 2,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	stmt = "alter vschema on test add vindex test_hash_id2 (id2) using hash"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ks, "test_hash_id2", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	if table, ok := vschema.Keyspaces[ks].Tables["test"]; ok {
-		if len(table.ColumnVindexes) != 3 {
-			t.Fatalf("table vindexes want 1 got %d", len(table.ColumnVindexes))
-		}
-		if table.ColumnVindexes[2].Name != "test_hash_id2" {
-			t.Fatalf("table vindexes didn't contain test_hash_id2")
-		}
-	} else {
-		t.Fatalf("table test not defined in vschema")
-	}
-
-	qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-	}
-	wantqr = &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-			buildVarCharRow("c1, c2", "test_lookup", "lookup", "from=c1,c2; table=test_lookup; to=keyspace_id", "test"),
-			buildVarCharRow("id2", "test_hash_id2", "hash", "", ""),
-		},
-		RowsAffected: 3,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	// drop one
-	stmt = "alter vschema on test drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	// wait for up to 50ms for it to disappear
-	deadline := time.Now().Add(50 * time.Millisecond)
-	for {
-		qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-		if err != nil {
-			t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-		}
-		wantqr = &sqltypes.Result{
-			Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-			Rows: [][]sqltypes.Value{
-				buildVarCharRow("id", "test_hash", "hash", "", ""),
-				buildVarCharRow("id2", "test_hash_id2", "hash", "", ""),
-			},
-			RowsAffected: 2,
-		}
-		if reflect.DeepEqual(qr, wantqr) {
-			break
-		}
-
-		if time.Now().After(deadline) {
-			t.Errorf("timed out waiting for test_lookup vindex to be removed")
-		}
-		time.Sleep(1 * time.Millisecond)
-	}
-
-	// use the newly created vindex on a new table
-	stmt = "alter vschema on test2 add vindex test_hash (id)"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ks, "test_hash", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	if table, ok := vschema.Keyspaces[ks].Tables["test2"]; ok {
-		if len(table.ColumnVindexes) != 1 {
-			t.Fatalf("table vindexes want 1 got %d", len(table.ColumnVindexes))
-		}
-		if table.ColumnVindexes[0].Name != "test_hash" {
-			t.Fatalf("table vindexes didn't contain test_hash")
-		}
-	} else {
-		t.Fatalf("table test2 not defined in vschema")
-	}
-
-	// create an identical vindex definition on a different table
-	stmt = "alter vschema on test2 add vindex test_lookup (c1,c2) using lookup with owner=`test`, from=`c1,c2`, table=test_lookup, to=keyspace_id"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ks, "test_lookup", vschemaUpdates, executor)
-	if vindex.Type != "lookup" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	if table, ok := vschema.Keyspaces[ks].Tables["test2"]; ok {
-		if len(table.ColumnVindexes) != 2 {
-			t.Fatalf("table vindexes want 1 got %d", len(table.ColumnVindexes))
-		}
-		if table.ColumnVindexes[1].Name != "test_lookup" {
-			t.Fatalf("table vindexes didn't contain test_lookup")
-		}
-	} else {
-		t.Fatalf("table test2 not defined in vschema")
-	}
-
-	qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test2", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test2: %v", err)
-	}
-	wantqr = &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-			buildVarCharRow("c1, c2", "test_lookup", "lookup", "from=c1,c2; table=test_lookup; to=keyspace_id", "test"),
-		},
-		RowsAffected: 2,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	stmt = "alter vschema on test2 add vindex nonexistent (c1,c2)"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex nonexistent does not exist in keyspace TestExecutor"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on test2 add vindex test_hash (c1,c2) using lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex test_hash defined with type hash not lookup"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on test2 add vindex test_lookup (c1,c2) using lookup with owner=xyz"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex test_lookup defined with owner test not xyz"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on test2 add vindex test_lookup (c1,c2) using lookup with owner=`test`, foo=bar"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex test_lookup defined with different parameters"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on nonexistent drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "table TestExecutor.nonexistent not defined in vschema"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on nonexistent drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", NewSafeSession(&vtgatepb.Session{TargetString: "InvalidKeyspace"}), stmt, nil)
-	wantErr = "no keyspace with name [InvalidKeyspace] found"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on nowhere.nohow drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "no keyspace with name [nowhere] found"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on test drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex test_lookup not defined in table TestExecutor.test"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	// no queries should have gone to any tablets
-	wantCount := []int64{0, 0, 0}
-	gotCount := []int64{
-		sbc1.ExecCount.Get(),
-		sbc2.ExecCount.Get(),
-		sbclookup.ExecCount.Get(),
-	}
-	if !reflect.DeepEqual(gotCount, wantCount) {
-		t.Errorf("Exec %s: %v, want %v", "", gotCount, wantCount)
-	}
-}
-
 func TestExecutorVindexDDLACL(t *testing.T) {
 	executor, _, _, _ := createExecutorEnv()
 	ks := "TestExecutor"
@@ -2361,4 +1938,222 @@ func TestOlapSelectDatabase(t *testing.T) {
 	err := executor.StreamExecute(context.Background(), "TestExecute", NewSafeSession(session), sql, nil, target, cb)
 	assert.NoError(t, err)
 	assert.True(t, cbInvoked)
+}
+
+func TestExecutorResultsExceeded(t *testing.T) {
+	save := *warnMemoryRows
+	*warnMemoryRows = 3
+	defer func() { *warnMemoryRows = save }()
+
+	executor, _, _, sbclookup := createExecutorEnv()
+	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master"})
+
+	initial := warnings.Counts()["ResultsExceeded"]
+
+	result1 := sqltypes.MakeTestResult(sqltypes.MakeTestFields("col", "int64"), "1")
+	result2 := sqltypes.MakeTestResult(sqltypes.MakeTestFields("col", "int64"), "1", "2", "3", "4")
+	sbclookup.SetResults([]*sqltypes.Result{result1, result2})
+
+	_, err := executor.Execute(context.Background(), "TestExecutorResultsExceeded", session, "select * from main1", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if got, want := warnings.Counts()["ResultsExceeded"], initial; got != want {
+		t.Errorf("warnings count: %v, want %v", got, want)
+	}
+
+	_, err = executor.Execute(context.Background(), "TestExecutorResultsExceeded", session, "select * from main1", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if got, want := warnings.Counts()["ResultsExceeded"], initial+1; got != want {
+		t.Errorf("warnings count: %v, want %v", got, want)
+	}
+}
+
+func TestExecutorClearsWarnings(t *testing.T) {
+	executor, _, _, _ := createExecutorEnv()
+	session := NewSafeSession(&vtgatepb.Session{
+		Warnings: []*querypb.QueryWarning{{Code: 234, Message: "oh noes"}},
+	})
+	_, err := executor.Execute(context.Background(), "TestExecute", session, "select 42", nil)
+	require.NoError(t, err)
+	require.Empty(t, session.Warnings)
+}
+
+func TestExecutorOtherRead(t *testing.T) {
+	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
+
+	type cnts struct {
+		Sbc1Cnt      int64
+		Sbc2Cnt      int64
+		SbcLookupCnt int64
+	}
+
+	tcs := []struct {
+		targetStr string
+
+		hasNoKeyspaceErr       bool
+		hasDestinationShardErr bool
+		wantCnts               cnts
+	}{
+		{
+			targetStr:        "",
+			hasNoKeyspaceErr: true,
+		},
+		{
+			targetStr:              "TestExecutor[-]",
+			hasDestinationShardErr: true,
+		},
+		{
+			targetStr: KsTestUnsharded,
+			wantCnts: cnts{
+				Sbc1Cnt:      0,
+				Sbc2Cnt:      0,
+				SbcLookupCnt: 1,
+			},
+		},
+		{
+			targetStr: "TestExecutor",
+			wantCnts: cnts{
+				Sbc1Cnt:      1,
+				Sbc2Cnt:      0,
+				SbcLookupCnt: 0,
+			},
+		},
+	}
+
+	stmts := []string{
+		"analyze table t1",
+		"describe select * from t1",
+		"explain select * from t1",
+		"do 1",
+	}
+
+	for _, stmt := range stmts {
+		for _, tc := range tcs {
+			sbc1.ExecCount.Set(0)
+			sbc2.ExecCount.Set(0)
+			sbclookup.ExecCount.Set(0)
+
+			_, err := executor.Execute(context.Background(), "TestExecute", NewSafeSession(&vtgatepb.Session{TargetString: tc.targetStr}), stmt, nil)
+			if tc.hasNoKeyspaceErr {
+				assert.EqualError(t, err, "keyspace not specified")
+			} else if tc.hasDestinationShardErr {
+				assert.Errorf(t, err, "Destination can only be a single shard for statement: %s, got: DestinationExactKeyRange(-)", stmt)
+			} else {
+				assert.NoError(t, err)
+			}
+
+			utils.MustMatch(t, tc.wantCnts, cnts{
+				Sbc1Cnt:      sbc1.ExecCount.Get(),
+				Sbc2Cnt:      sbc2.ExecCount.Get(),
+				SbcLookupCnt: sbclookup.ExecCount.Get(),
+			}, "count did not match")
+		}
+	}
+}
+
+func TestExecutorExplain(t *testing.T) {
+	executor, _, _, _ := createExecutorEnv()
+	executor.normalize = true
+	logChan := QueryLogger.Subscribe("Test")
+	defer QueryLogger.Unsubscribe(logChan)
+
+	bindVars := map[string]*querypb.BindVariable{}
+	result, err := executorExec(executor, "explain format = vitess select * from user", bindVars)
+	require.NoError(t, err)
+
+	require.Equal(t,
+		`[[VARCHAR("Route") VARCHAR("SelectScatter") VARCHAR("TestExecutor") VARCHAR("") VARCHAR("UNKNOWN") VARCHAR("select * from user")]]`,
+		fmt.Sprintf("%v", result.Rows))
+
+	result, err = executorExec(executor, "explain format = vitess select 42", bindVars)
+	require.NoError(t, err)
+	expected :=
+		`[[VARCHAR("Projection") VARCHAR("") VARCHAR("") VARCHAR("") VARCHAR("UNKNOWN") VARCHAR("")] ` +
+			`[VARCHAR("└─ SingleRow") VARCHAR("") VARCHAR("") VARCHAR("") VARCHAR("UNKNOWN") VARCHAR("")]]`
+	require.Equal(t,
+		`[[VARCHAR("Projection") VARCHAR("") VARCHAR("") VARCHAR("") VARCHAR("UNKNOWN") VARCHAR("")] `+
+			`[VARCHAR("└─ SingleRow") VARCHAR("") VARCHAR("") VARCHAR("") VARCHAR("UNKNOWN") VARCHAR("")]]`,
+		expected,
+		fmt.Sprintf("%v", result.Rows), fmt.Sprintf("%v", result.Rows))
+}
+
+func TestExecutorOtherAdmin(t *testing.T) {
+	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
+
+	type cnts struct {
+		Sbc1Cnt      int64
+		Sbc2Cnt      int64
+		SbcLookupCnt int64
+	}
+
+	tcs := []struct {
+		targetStr string
+
+		hasNoKeyspaceErr       bool
+		hasDestinationShardErr bool
+		wantCnts               cnts
+	}{
+		{
+			targetStr:        "",
+			hasNoKeyspaceErr: true,
+		},
+		{
+			targetStr:              "TestExecutor[-]",
+			hasDestinationShardErr: true,
+		},
+		{
+			targetStr: KsTestUnsharded,
+			wantCnts: cnts{
+				Sbc1Cnt:      0,
+				Sbc2Cnt:      0,
+				SbcLookupCnt: 1,
+			},
+		},
+		{
+			targetStr: "TestExecutor",
+			wantCnts: cnts{
+				Sbc1Cnt:      1,
+				Sbc2Cnt:      0,
+				SbcLookupCnt: 0,
+			},
+		},
+	}
+
+	stmts := []string{
+		"repair table t1",
+		"optimize table t1",
+	}
+
+	for _, stmt := range stmts {
+		for _, tc := range tcs {
+			sbc1.ExecCount.Set(0)
+			sbc2.ExecCount.Set(0)
+			sbclookup.ExecCount.Set(0)
+
+			_, err := executor.Execute(context.Background(), "TestExecute", NewSafeSession(&vtgatepb.Session{TargetString: tc.targetStr}), stmt, nil)
+			if tc.hasNoKeyspaceErr {
+				assert.Error(t, err, errNoKeyspace)
+			} else if tc.hasDestinationShardErr {
+				assert.Errorf(t, err, "Destination can only be a single shard for statement: %s, got: DestinationExactKeyRange(-)", stmt)
+			} else {
+				assert.NoError(t, err)
+			}
+
+			diff := cmp.Diff(tc.wantCnts, cnts{
+				Sbc1Cnt:      sbc1.ExecCount.Get(),
+				Sbc2Cnt:      sbc2.ExecCount.Get(),
+				SbcLookupCnt: sbclookup.ExecCount.Get(),
+			})
+			if diff != "" {
+				t.Errorf("stmt: %s\ntc: %+v\n-want,+got:\n%s", stmt, tc, diff)
+			}
+		}
+	}
+}
+
+func makeComments(text string) sqlparser.MarginComments {
+	return sqlparser.MarginComments{Trailing: text}
 }
